third_party/blink/web_tests/platform/win/external/wpt/layout-instability/input-timestamp-expected.txt

This is a testharness.js-based test.
FAIL Layout shift right after user input is observable via PerformanceObserver. assert_equals: expected false but got true
Harness: the test ran to completion.
